[Decomp][Spell] Cohesion split: extract tail handlers to SpellEffectTail.cpp
Move the final contiguous block of effect handlers (EffectDispelMechanic
through EffectCreateTamedPet, the end of the file) out of SpellEffects.cpp
into a new SpellEffectTail.cpp translation unit. Covers totems, durability,
WMO, kill-credit, quest, rune/spec and other miscellaneous tail effects.

Pure file-level cohesion split: declarations stay in Spell.h, the
SpellEffects[] dispatch table stays in SpellEffects.cpp and resolves the
moved member methods by address at link time. Byte-identical move, no
behaviour change. SpellEffects.cpp 6027 -> 4942.
