+++ src/game/WorldHandlers/SpellEffects.cpp
@@ -4940,1088 +4940,3 @@ void Spell::EffectPlayerPull(SpellEffectEntry const* effect)
     float speed = m_spellInfo->speed ? m_spellInfo->speed : 27.0f;
     unitTarget->GetMotionMaster()->MoveJump(x, y, z, speed, 2.5f);
 }
-
-/**
- * @brief Removes auras from the target that match the specified mechanic.
- *
- * @param effect The effect index containing the mechanic id.
- */
-void Spell::EffectDispelMechanic(SpellEffectEntry const* effect)
-{
-    if (!unitTarget)
-    {
-        return;
-    }
-
-    uint32 mechanic = effect->EffectMiscValue;
-
-    Unit::SpellAuraHolderMap& Auras = unitTarget->GetSpellAuraHolderMap();
-    for (Unit::SpellAuraHolderMap::iterator iter = Auras.begin(), next; iter != Auras.end(); iter = next)
-    {
-        next = iter;
-        ++next;
-        SpellEntry const* spell = iter->second->GetSpellProto();
-        if (iter->second->HasMechanic(mechanic))
-        {
-            unitTarget->RemoveAurasDueToSpell(spell->Id);
-            if (Auras.empty())
-            {
-                break;
-            }
-            else
-            {
-                next = Auras.begin();
-            }
-        }
-    }
-}
-
-/**
- * @brief Restores the caster's dead pet and revives it with percentage-based health.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectSummonDeadPet(SpellEffectEntry const* /*effect*/)
-{
-    if (m_caster->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-    Player* _player = (Player*)m_caster;
-    Pet* pet = _player->GetPet();
-    if (!pet)
-    {
-        return;
-    }
-    if (pet->IsAlive())
-    {
-        return;
-    }
-
-    if (_player->GetDistance(pet) >= 2.0f)
-    {
-        float px, py, pz;
-        m_caster->GetClosePoint(px, py, pz, pet->GetObjectBoundingRadius());
-        ((Unit*)pet)->NearTeleportTo(px, py, pz, -m_caster->GetOrientation());
-    }
-
-    pet->SetUInt32Value(UNIT_DYNAMIC_FLAGS, UNIT_DYNFLAG_NONE);
-    pet->RemoveFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_SKINNABLE);
-    pet->SetDeathState(ALIVE);
-    pet->clearUnitState(UNIT_STAT_ALL_STATE);
-    pet->SetHealth(uint32(pet->GetMaxHealth() * (float(damage) / 100)));
-
-    pet->AIM_Initialize();
-
-    // _player->PetSpellInitialize(); -- action bar not removed at death and not required send at revive
-    pet->SavePetToDB(PET_SAVE_AS_CURRENT);
-}
-
-void Spell::EffectSummonAllTotems(SpellEffectEntry const* effect)
-{
-    if (m_caster->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    int32 start_button = ACTION_BUTTON_SHAMAN_TOTEMS_BAR + effect->EffectMiscValue;
-    int32 amount_buttons = effect->EffectMiscValueB;
-
-    for (int32 slot = 0; slot < amount_buttons; ++slot)
-        if (ActionButton const* actionButton = ((Player*)m_caster)->GetActionButton(start_button + slot))
-            if (actionButton->GetType() == ACTION_BUTTON_SPELL)
-                if (uint32 spell_id = actionButton->GetAction())
-                {
-                    m_caster->CastSpell(unitTarget, spell_id, true);
-                }
-}
-
-/**
- * @brief Unsummons all totems currently owned by the caster.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectDestroyAllTotems(SpellEffectEntry const* /*effect*/)
-{
-    int32 mana = 0;
-    for (int slot = 0;  slot < MAX_TOTEM_SLOT; ++slot)
-    {
-        if (Totem* totem = m_caster->GetTotem(TotemSlot(slot)))
-        {
-            if (damage)
-            {
-                uint32 spell_id = totem->GetUInt32Value(UNIT_CREATED_BY_SPELL);
-                if (SpellEntry const* spellInfo = sSpellStore.LookupEntry(spell_id))
-                {
-                    uint32 manacost = m_caster->GetCreateMana() * spellInfo->GetManaCostPercentage() / 100;
-                    mana += manacost * damage / 100;
-                }
-            }
-            totem->UnSummon();
-        }
-    }
-
-    if (mana)
-    {
-        m_caster->CastCustomSpell(m_caster, 39104, &mana, NULL, NULL, true);
-    }
-}
-
-void Spell::EffectBreakPlayerTargeting (SpellEffectEntry const* /*effect*/)
-{
-    if (!unitTarget)
-    {
-        return;
-    }
-
-    WorldPacket data(SMSG_CLEAR_TARGET, 8);
-    data << unitTarget->GetObjectGuid();
-    unitTarget->SendMessageToSet(&data, false);
-}
-
-/**
- * @brief Removes a fixed number of durability points from one or more player items.
- *
- * @param effect The effect index containing the inventory slot selector.
- */
-void Spell::EffectDurabilityDamage(SpellEffectEntry const* effect)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    int32 slot = effect->EffectMiscValue;
-
-    // FIXME: some spells effects have value -1/-2
-    // Possibly its mean -1 all player equipped items and -2 all items
-    if (slot < 0)
-    {
-        ((Player*)unitTarget)->DurabilityPointsLossAll(damage, (slot < -1));
-        return;
-    }
-
-    // invalid slot value
-    if (slot >= INVENTORY_SLOT_BAG_END)
-    {
-        return;
-    }
-
-    if (Item* item = ((Player*)unitTarget)->GetItemByPos(INVENTORY_SLOT_BAG_0, slot))
-    {
-        ((Player*)unitTarget)->DurabilityPointsLoss(item, damage);
-    }
-}
-
-/**
- * @brief Removes a percentage of durability from one or more player items.
- *
- * @param effect The effect index containing the inventory slot selector.
- */
-void Spell::EffectDurabilityDamagePCT(SpellEffectEntry const* effect)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    int32 slot = effect->EffectMiscValue;
-
-    // FIXME: some spells effects have value -1/-2
-    // Possibly its mean -1 all player equipped items and -2 all items
-    if (slot < 0)
-    {
-        ((Player*)unitTarget)->DurabilityLossAll(double(damage) / 100.0f, (slot < -1));
-        return;
-    }
-
-    // invalid slot value
-    if (slot >= INVENTORY_SLOT_BAG_END)
-    {
-        return;
-    }
-
-    if (damage <= 0)
-    {
-        return;
-    }
-
-    if (Item* item = ((Player*)unitTarget)->GetItemByPos(INVENTORY_SLOT_BAG_0, slot))
-    {
-        ((Player*)unitTarget)->DurabilityLoss(item, double(damage) / 100.0f);
-    }
-}
-
-/**
- * @brief Modifies the caster's threat on the target by a percentage.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectModifyThreatPercent(SpellEffectEntry const* /*effect*/)
-{
-    if (!unitTarget)
-    {
-        return;
-    }
-
-    unitTarget->GetThreatManager().modifyThreatPercent(m_caster, damage);
-}
-
-/**
- * @brief Summons a transmitted game object such as fishing nodes, rituals, or spell casters.
- *
- * @param effect The effect index containing the game object entry.
- */
-void Spell::EffectTransmitted(SpellEffectEntry const* effect)
-{
-    uint32 name_id = effect->EffectMiscValue;
-
-    switch (m_spellInfo->Id)
-    {
-        case 29886: // Create Soulwell
-            if (m_caster->HasAura(18692))
-            {
-                name_id = 183510;
-            }
-            else if (m_caster->HasAura(18693))
-            {
-                name_id = 183511;
-            }
-            break;
-        default:
-            break;
-    }
-
-    GameObjectInfo const* goinfo = ObjectMgr::GetGameObjectInfo(name_id);
-
-    if (!goinfo)
-    {
-        sLog.outErrorDb("Gameobject (Entry: %u) not exist and not created at spell (ID: %u) cast", name_id, m_spellInfo->Id);
-        return;
-    }
-
-    float fx, fy, fz;
-
-    if (m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION)
-    {
-        m_targets.getDestination(fx, fy, fz);
-    }
-    // FIXME: this can be better check for most objects but still hack
-    else if (effect->GetRadiusIndex() && m_spellInfo->speed == 0)
-    {
-        float dis = GetSpellRadius(sSpellRadiusStore.LookupEntry(effect->GetRadiusIndex()));
-        m_caster->GetClosePoint(fx, fy, fz, DEFAULT_WORLD_OBJECT_SIZE, dis);
-    }
-    else
-    {
-        float min_dis = GetSpellMinRange(sSpellRangeStore.LookupEntry(m_spellInfo->rangeIndex));
-        float max_dis = GetSpellMaxRange(sSpellRangeStore.LookupEntry(m_spellInfo->rangeIndex));
-        float dis = rand_norm_f() * (max_dis - min_dis) + min_dis;
-
-        // special code for fishing bobber (TARGET_SELF_FISHING), should not try to avoid objects
-        // nor try to find ground level, but randomly vary in angle
-        if (goinfo->type == GAMEOBJECT_TYPE_FISHINGNODE)
-        {
-            // calculate angle variation for roughly equal dimensions of target area
-            float max_angle = (max_dis - min_dis) / (max_dis + m_caster->GetObjectBoundingRadius());
-            float angle_offset = max_angle * (rand_norm_f() - 0.5f);
-            m_caster->GetNearPoint2D(fx, fy, dis + m_caster->GetObjectBoundingRadius(), m_caster->GetOrientation() + angle_offset);
-
-            GridMapLiquidData liqData;
-            if (!m_caster->GetTerrain()->IsInWater(fx, fy, m_caster->GetPositionZ() + 1.f, &liqData))
-            {
-                SendCastResult(SPELL_FAILED_NOT_FISHABLE);
-                SendChannelUpdate(0);
-                return;
-            }
-
-            fz = liqData.level;
-            // finally, check LoS
-            if (!m_caster->IsWithinLOS(fx, fy, fz))
-            {
-                SendCastResult(SPELL_FAILED_LINE_OF_SIGHT);
-                SendChannelUpdate(0);
-                return;
-            }
-        }
-        else
-        {
-            m_caster->GetClosePoint(fx, fy, fz, DEFAULT_WORLD_OBJECT_SIZE, dis);
-        }
-    }
-
-    Map* cMap = m_caster->GetMap();
-
-    // if gameobject is summoning object, it should be spawned right on caster's position
-    if (goinfo->type == GAMEOBJECT_TYPE_SUMMONING_RITUAL)
-    {
-        m_caster->GetPosition(fx, fy, fz);
-    }
-
-    GameObject* pGameObj = new GameObject;
-
-    if (!pGameObj->Create(cMap->GenerateLocalLowGuid(HIGHGUID_GAMEOBJECT), name_id, cMap,
-                          m_caster->GetPhaseMask(), fx, fy, fz, m_caster->GetOrientation()))
-    {
-        delete pGameObj;
-        return;
-    }
-
-    int32 duration = m_duration;
-
-    switch (goinfo->type)
-    {
-        case GAMEOBJECT_TYPE_FISHINGNODE:
-        {
-            m_caster->SetChannelObjectGuid(pGameObj->GetObjectGuid());
-            m_caster->AddGameObject(pGameObj);              // will removed at spell cancel
-
-            // end time of range when possible catch fish (FISHING_BOBBER_READY_TIME..GetDuration(m_spellInfo))
-            // start time == fish-FISHING_BOBBER_READY_TIME (0..GetDuration(m_spellInfo)-FISHING_BOBBER_READY_TIME)
-            int32 lastSec = 0;
-            switch (urand(0, 3))
-            {
-                case 0: lastSec =  3; break;
-                case 1: lastSec =  7; break;
-                case 2: lastSec = 13; break;
-                case 3: lastSec = 17; break;
-            }
-
-            duration = duration - lastSec * IN_MILLISECONDS + FISHING_BOBBER_READY_TIME * IN_MILLISECONDS;
-            break;
-        }
-        case GAMEOBJECT_TYPE_SUMMONING_RITUAL:
-        {
-            if (m_caster->GetTypeId() == TYPEID_PLAYER)
-            {
-                pGameObj->AddUniqueUse((Player*)m_caster);
-                m_caster->AddGameObject(pGameObj);          // will removed at spell cancel
-            }
-            break;
-        }
-        case GAMEOBJECT_TYPE_SPELLCASTER:
-        {
-            m_caster->AddGameObject(pGameObj);
-            break;
-        }
-        case GAMEOBJECT_TYPE_FISHINGHOLE:
-        case GAMEOBJECT_TYPE_CHEST:
-        default:
-            break;
-    }
-
-    pGameObj->SetRespawnTime(duration > 0 ? duration / IN_MILLISECONDS : 0);
-
-    pGameObj->SetOwnerGuid(m_caster->GetObjectGuid());
-
-    pGameObj->SetUInt32Value(GAMEOBJECT_LEVEL, m_caster->getLevel());
-    pGameObj->SetSpellId(m_spellInfo->Id);
-
-    DEBUG_LOG("AddObject at SpellEfects.cpp EffectTransmitted");
-    // m_caster->AddGameObject(pGameObj);
-    // m_ObjToDel.push_back(pGameObj);
-
-    cMap->Add(pGameObj);
-
-    pGameObj->SummonLinkedTrapIfAny();
-
-    if (m_caster->GetTypeId() == TYPEID_UNIT && ((Creature*)m_caster)->AI())
-    {
-        ((Creature*)m_caster)->AI()->JustSummoned(pGameObj);
-    }
-    if (m_originalCaster && m_originalCaster != m_caster && m_originalCaster->GetTypeId() == TYPEID_UNIT && ((Creature*)m_originalCaster)->AI())
-    {
-        ((Creature*)m_originalCaster)->AI()->JustSummoned(pGameObj);
-    }
-}
-
-void Spell::EffectProspecting(SpellEffectEntry const* /*effect*/)
-{
-    if (m_caster->GetTypeId() != TYPEID_PLAYER || !itemTarget)
-    {
-        return;
-    }
-
-    Player* p_caster = (Player*)m_caster;
-
-    if (sWorld.getConfig(CONFIG_BOOL_SKILL_PROSPECTING))
-    {
-        uint32 SkillValue = p_caster->GetPureSkillValue(SKILL_JEWELCRAFTING);
-        uint32 reqSkillValue = itemTarget->GetProto()->RequiredSkillRank;
-        p_caster->UpdateGatherSkill(SKILL_JEWELCRAFTING, SkillValue, reqSkillValue);
-    }
-
-    ((Player*)m_caster)->SendLoot(itemTarget->GetObjectGuid(), LOOT_PROSPECTING);
-}
-
-void Spell::EffectMilling(SpellEffectEntry const* /*effect*/)
-{
-    if (m_caster->GetTypeId() != TYPEID_PLAYER || !itemTarget)
-    {
-        return;
-    }
-
-    Player* p_caster = (Player*)m_caster;
-
-    if (sWorld.getConfig(CONFIG_BOOL_SKILL_MILLING))
-    {
-        uint32 SkillValue = p_caster->GetPureSkillValue(SKILL_INSCRIPTION);
-        uint32 reqSkillValue = itemTarget->GetProto()->RequiredSkillRank;
-        p_caster->UpdateGatherSkill(SKILL_INSCRIPTION, SkillValue, reqSkillValue);
-    }
-
-    ((Player*)m_caster)->SendLoot(itemTarget->GetObjectGuid(), LOOT_MILLING);
-}
-
-void Spell::EffectSkill(SpellEffectEntry const* /*effect*/)
-{
-    DEBUG_LOG("WORLD: SkillEFFECT");
-}
-
-/**
- * @brief Fully resurrects a dead player target as part of a spirit heal effect.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectSpiritHeal(SpellEffectEntry const* /*effect*/)
-{
-    // TODO player can't see the heal-animation - he should respawn some ticks later
-    if (!unitTarget || unitTarget->IsAlive())
-    {
-        return;
-    }
-    if (unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-    if (!unitTarget->IsInWorld())
-    {
-        return;
-    }
-    if (m_spellInfo->Id == 22012 && !unitTarget->HasAura(2584))
-    {
-        return;
-    }
-
-    ((Player*)unitTarget)->ResurrectPlayer(1.0f);
-    ((Player*)unitTarget)->SpawnCorpseBones();
-}
-
-// remove insignia spell effect
-void Spell::EffectSkinPlayerCorpse(SpellEffectEntry const* /*effect*/)
-{
-    DEBUG_LOG("Effect: SkinPlayerCorpse");
-    if ((m_caster->GetTypeId() != TYPEID_PLAYER) || (unitTarget->GetTypeId() != TYPEID_PLAYER) || (unitTarget->IsAlive()))
-    {
-        return;
-    }
-
-    ((Player*)unitTarget)->RemovedInsignia((Player*)m_caster);
-}
-
-void Spell::EffectStealBeneficialBuff(SpellEffectEntry const* effect)
-{
-    DEBUG_LOG("Effect: StealBeneficialBuff");
-
-    if (!unitTarget || unitTarget == m_caster)              // can't steal from self
-    {
-        return;
-    }
-
-    typedef std::vector<SpellAuraHolder*> StealList;
-    StealList steal_list;
-    // Create dispel mask by dispel type
-    uint32 dispelMask  = GetDispellMask( DispelType(effect->EffectMiscValue) );
-    Unit::SpellAuraHolderMap const& auras = unitTarget->GetSpellAuraHolderMap();
-    for (Unit::SpellAuraHolderMap::const_iterator itr = auras.begin(); itr != auras.end(); ++itr)
-    {
-        SpellAuraHolder *holder = itr->second;
-        if (holder && (1<<holder->GetSpellProto()->GetDispel()) & dispelMask)
-        {
-            // Need check for passive? this
-            if (holder->IsPositive() && !holder->IsPassive() && !holder->GetSpellProto()->HasAttribute(SPELL_ATTR_EX4_NOT_STEALABLE))
-            {
-                steal_list.push_back(holder);
-            }
-        }
-    }
-    // Ok if exist some buffs for dispel try dispel it
-    if (!steal_list.empty())
-    {
-        typedef std::list < std::pair<uint32, ObjectGuid> > SuccessList;
-        SuccessList success_list;
-        int32 list_size = steal_list.size();
-        // Dispell N = damage buffs (or while exist buffs for dispel)
-        for (int32 count = 0; count < damage && list_size > 0; ++count)
-        {
-            // Random select buff for dispel
-            SpellAuraHolder* holder = steal_list[urand(0, list_size - 1)];
-            // Not use chance for steal
-            // TODO possible need do it
-            success_list.push_back(SuccessList::value_type(holder->GetId(), holder->GetCasterGuid()));
-
-            // Remove buff from list for prevent doubles
-            for (StealList::iterator j = steal_list.begin(); j != steal_list.end();)
-            {
-                SpellAuraHolder* stealed = *j;
-                if (stealed->GetId() == holder->GetId() && stealed->GetCasterGuid() == holder->GetCasterGuid())
-                {
-                    j = steal_list.erase(j);
-                    --list_size;
-                }
-                else
-                {
-                    ++j;
-                }
-            }
-        }
-        // Really try steal and send log
-        if (!success_list.empty())
-        {
-            int32 count = success_list.size();
-            WorldPacket data(SMSG_SPELLSTEALLOG, 8 + 8 + 4 + 1 + 4 + count * 5);
-            data << unitTarget->GetPackGUID();       // Victim GUID
-            data << m_caster->GetPackGUID();         // Caster GUID
-            data << uint32(m_spellInfo->Id);         // Dispell spell id
-            data << uint8(0);                        // not used
-            data << uint32(count);                   // count
-            for (SuccessList::iterator j = success_list.begin(); j != success_list.end(); ++j)
-            {
-                SpellEntry const* spellInfo = sSpellStore.LookupEntry(j->first);
-                data << uint32(spellInfo->Id);       // Spell Id
-                data << uint8(0);                    // 0 - steals !=0 transfers
-                unitTarget->RemoveAurasDueToSpellBySteal(spellInfo->Id, j->second, m_caster);
-            }
-            m_caster->SendMessageToSet(&data, true);
-        }
-    }
-}
-
-void Spell::EffectWMODamage(SpellEffectEntry const* effect)
-{
-    DEBUG_LOG("Effect: WMODamage");
-
-    if (!gameObjTarget)
-    {
-        return;
-    }
-
-    if (gameObjTarget->GetGoType() != GAMEOBJECT_TYPE_DESTRUCTIBLE_BUILDING)
-    {
-        sLog.outError("Spell::EffectWMODamage called without valid targets. Spell Id %u", m_spellInfo->Id);
-        return;
-    }
-
-    if (!gameObjTarget->GetHealth())
-    {
-        return;
-    }
-
-    Unit* caster = GetAffectiveCaster();
-    if (!caster)
-    {
-        return;
-    }
-
-    DEBUG_LOG("Spell::EffectWMODamage, spell Id %u, go entry %u, damage %u", m_spellInfo->Id, gameObjTarget->GetEntry(), uint32(damage));
-    gameObjTarget->DealGameObjectDamage(uint32(damage), m_spellInfo->Id, caster);
-}
-
-void Spell::EffectWMORepair(SpellEffectEntry const* effect)
-{
-    DEBUG_LOG("Effect: WMORepair");
-
-    if (!gameObjTarget)
-    {
-        return;
-    }
-
-    if (gameObjTarget->GetGoType() != GAMEOBJECT_TYPE_DESTRUCTIBLE_BUILDING)
-    {
-        sLog.outError("Spell::EffectWMORepair called without valid targets. Spell Id %u", m_spellInfo->Id);
-        return;
-    }
-
-    Unit* caster = GetAffectiveCaster();
-    if (!caster)
-    {
-        return;
-    }
-
-    DEBUG_LOG("Spell::EffectWMORepair, spell Id %u, go entry %u", m_spellInfo->Id, gameObjTarget->GetEntry());
-    gameObjTarget->RebuildGameObject(m_spellInfo->Id, caster);
-}
-
-void Spell::EffectWMOChange(SpellEffectEntry const* effect)
-{
-    DEBUG_LOG("Effect: WMOChange");
-
-    if (!gameObjTarget)
-    {
-        return;
-    }
-
-    if (gameObjTarget->GetGoType() != GAMEOBJECT_TYPE_DESTRUCTIBLE_BUILDING)
-    {
-        sLog.outError("Spell::EffectWMOChange called without valid targets. Spell Id %u", m_spellInfo->Id);
-        return;
-    }
-
-    DEBUG_LOG("Spell::EffectWMOChange, spell Id %u, object %u, misc-value %u", m_spellInfo->Id, gameObjTarget->GetEntry(), effect->EffectMiscValue);
-
-    Unit* caster = GetAffectiveCaster();
-    if (!caster)
-    {
-        return;
-    }
-
-    switch (effect->EffectMiscValue)
-    {
-        case 0:                                             // Set to full health
-            gameObjTarget->ForceGameObjectHealth(gameObjTarget->GetMaxHealth(), caster);
-            break;
-        case 1:                                             // Set to damaged
-            gameObjTarget->ForceGameObjectHealth(gameObjTarget->GetGOInfo()->destructibleBuilding.damagedNumHits, caster);
-            break;
-        case 2:                                             // Set to destroyed
-            gameObjTarget->ForceGameObjectHealth(-int32(gameObjTarget->GetHealth()), caster);
-            break;
-        case 3:                                             // Set to rebuilding
-            gameObjTarget->ForceGameObjectHealth(0, caster);
-            break;
-        default:
-            sLog.outError("Spell::EffectWMOChange, spell Id %u with undefined change value %u", m_spellInfo->Id, effect->EffectMiscValue);
-            break;
-    }
-}
-
-void Spell::EffectKillCreditPersonal(SpellEffectEntry const* effect)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    ((Player*)unitTarget)->KilledMonsterCredit(effect->EffectMiscValue);
-}
-
-void Spell::EffectKillCreditGroup(SpellEffectEntry const* effect)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    ((Player*)unitTarget)->RewardPlayerAndGroupAtEvent(effect->EffectMiscValue, unitTarget);
-}
-
-void Spell::EffectQuestFail(SpellEffectEntry const* effect)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    ((Player*)unitTarget)->FailQuest(effect->EffectMiscValue);
-}
-
-void Spell::EffectActivateRune(SpellEffectEntry const* effect)
-{
-    if (m_caster->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    Player* plr = (Player*)m_caster;
-
-    if (plr->getClass() != CLASS_DEATH_KNIGHT)
-    {
-        return;
-    }
-
-    int32 count = damage;                                   // max amount of reset runes
-
-    plr->ResyncRunes();
-}
-
-void Spell::EffectTitanGrip(SpellEffectEntry const* effect)
-{
-    // Make sure "Titan's Grip" (49152) penalty spell does not silently change
-    if (effect->EffectMiscValue != 49152)
-    {
-        sLog.outError("Spell::EffectTitanGrip: Spell %u has unexpected EffectMiscValue '%u'", m_spellInfo->Id, effect->EffectMiscValue);
-    }
-    if (unitTarget && unitTarget->GetTypeId() == TYPEID_PLAYER)
-    {
-        Player* plr = (Player*)m_caster;
-        plr->SetCanTitanGrip(true);
-        if (plr->HasTwoHandWeaponInOneHand() && !plr->HasAura(49152))
-        {
-            plr->CastSpell(plr, 49152, true);
-        }
-    }
-}
-
-void Spell::EffectRenamePet(SpellEffectEntry const* /*effect*/)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_UNIT ||
-            !((Creature*)unitTarget)->IsPet() || ((Pet*)unitTarget)->getPetType() != HUNTER_PET)
-        return;
-
-    unitTarget->RemoveByteFlag(UNIT_FIELD_BYTES_2, 2, UNIT_CAN_BE_RENAMED);
-}
-
-void Spell::EffectPlaySound(SpellEffectEntry const* effect)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    uint32 soundId = effect->EffectMiscValue;
-    if (!sSoundEntriesStore.LookupEntry(soundId))
-    {
-        sLog.outError("EffectPlaySound: Sound (Id: %u) in spell %u does not exist.", soundId, m_spellInfo->Id);
-        return;
-    }
-
-    unitTarget->PlayDirectSound(soundId, (Player*)unitTarget);
-}
-
-void Spell::EffectPlayMusic(SpellEffectEntry const* effect)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    uint32 soundId = effect->EffectMiscValue;
-    if (!sSoundEntriesStore.LookupEntry(soundId))
-    {
-        sLog.outError("EffectPlayMusic: Sound (Id: %u) in spell %u does not exist.", soundId, m_spellInfo->Id);
-        return;
-    }
-
-    m_caster->PlayMusic(soundId, (Player*)unitTarget);
-}
-
-void Spell::EffectSpecCount(SpellEffectEntry const* /*effect*/)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    ((Player*)unitTarget)->UpdateSpecCount(damage);
-}
-
-void Spell::EffectActivateSpec(SpellEffectEntry const* /*effect*/)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    uint32 spec = damage - 1;
-
-    ((Player*)unitTarget)->ActivateSpec(spec);
-}
-
-/**
- * @brief Sets the player's homebind location to the current position.
- *
- * @param effect The bind effect index.
- */
-void Spell::EffectBind(SpellEffectEntry const* effect)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    Player* player = (Player*)unitTarget;
-
-    uint32 area_id = uint32(effect->EffectMiscValue);
-    WorldLocation loc;
-    if (effect->EffectImplicitTargetA == TARGET_TABLE_X_Y_Z_COORDINATES ||
-        effect->EffectImplicitTargetB == TARGET_TABLE_X_Y_Z_COORDINATES)
-    {
-        SpellTargetPosition const* st = sSpellMgr.GetSpellTargetPosition(m_spellInfo->Id);
-        if (!st)
-        {
-            sLog.outError("Spell::EffectBind - unknown Teleport coordinates for spell ID %u", m_spellInfo->Id);
-            return;
-        }
-
-        loc.mapid       = st->target_mapId;
-        loc.coord_x     = st->target_X;
-        loc.coord_y     = st->target_Y;
-        loc.coord_z     = st->target_Z;
-        loc.orientation = st->target_Orientation;
-        if (!area_id)
-        {
-            area_id = sTerrainMgr.GetAreaId(loc.mapid, loc.coord_x, loc.coord_y, loc.coord_z);
-        }
-    }
-    else
-    {
-        player->GetPosition(loc);
-        if (!area_id)
-        {
-            area_id = player->GetAreaId();
-        }
-    }
-
-    player->SetHomebindToLocation(loc, area_id);
-
-    // binding
-    WorldPacket data(SMSG_BINDPOINTUPDATE, (4 + 4 + 4 + 4 + 4));
-    data << float(loc.coord_x);
-    data << float(loc.coord_y);
-    data << float(loc.coord_z);
-    data << uint32(loc.mapid);
-    data << uint32(area_id);
-    player->SendDirectMessage(&data);
-
-    DEBUG_LOG("New Home Position for %s: XYZ: %f %f %f on Map %u", player->GetGuidStr().c_str(), loc.coord_x, loc.coord_y, loc.coord_z, loc.mapid);
-
-    // zone update
-    data.Initialize(SMSG_PLAYERBOUND, 8 + 4);
-    data << m_caster->GetObjectGuid();
-    data << uint32(area_id);
-    player->SendDirectMessage(&data);
-}
-
-void Spell::EffectRestoreItemCharges(SpellEffectEntry const* effect)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    Player* player = (Player*)unitTarget;
-
-    ItemPrototype const* itemProto = ObjectMgr::GetItemPrototype(effect->EffectItemType);
-    if (!itemProto)
-    {
-        return;
-    }
-
-    // In case item from limited category recharge any from category, is this valid checked early in spell checks
-    Item* item;
-    if (itemProto->ItemLimitCategory)
-    {
-        item = ((Player*)unitTarget)->GetItemByLimitedCategory(itemProto->ItemLimitCategory);
-    }
-    else
-    {
-        item = player->GetItemByEntry(effect->EffectItemType);
-    }
-
-    if (!item)
-    {
-        return;
-    }
-
-    item->RestoreCharges();
-}
-
-/**
- * @brief Sends a battleground player target to its graveyard.
- *
- * @param effect The teleport effect index.
- */
-void Spell::EffectRedirectThreat(SpellEffectEntry const* effect)
-{
-    if (!unitTarget)
-    {
-        return;
-    }
-
-    if (m_spellInfo->Id == 59665)                           // Vigilance
-        if (Aura* glyph = unitTarget->GetDummyAura(63326))  // Glyph of Vigilance
-        {
-            damage += glyph->GetModifier()->m_amount;
-        }
-
-    m_caster->GetHostileRefManager().SetThreatRedirection(unitTarget->GetObjectGuid(), uint32(damage));
-}
-
-void Spell::EffectTeachTaxiNode(SpellEffectEntry const* effect)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    Player* player = (Player*)unitTarget;
-
-    uint32 taxiNodeId = effect->EffectMiscValue;
-    if (!sTaxiNodesStore.LookupEntry(taxiNodeId))
-    {
-        return;
-    }
-
-    if (player->m_taxi.SetTaximaskNode(taxiNodeId))
-    {
-        WorldPacket data(SMSG_NEW_TAXI_PATH, 0);
-        player->SendDirectMessage(&data);
-
-        data.Initialize(SMSG_TAXINODE_STATUS, 9);
-        data << m_caster->GetObjectGuid();
-        data << uint8(1);
-        player->SendDirectMessage(&data);
-    }
-}
-
-void Spell::EffectQuestOffer(SpellEffectEntry const* effect)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    if (Quest const* quest = sObjectMgr.GetQuestTemplate(effect->EffectMiscValue))
-    {
-        Player* player = (Player*)unitTarget;
-
-        if (player->CanTakeQuest(quest, false))
-        {
-            player->PlayerTalkClass->SendQuestGiverQuestDetails(quest, player->GetObjectGuid(), true);
-        }
-    }
-}
-
-void Spell::EffectCancelAura(SpellEffectEntry const* effect)
-{
-    if (!unitTarget)
-    {
-        return;
-    }
-
-    uint32 spellId = effect->EffectTriggerSpell;
-
-    if (!sSpellStore.LookupEntry(spellId))
-    {
-        sLog.outError("Spell::EffectCancelAura: spell %u doesn't exist", spellId);
-        return;
-    }
-
-    unitTarget->RemoveAurasDueToSpell(spellId);
-}
-
-void Spell::EffectKnockBackFromPosition(SpellEffectEntry const* effect)
-{
-    if (!unitTarget)
-    {
-        return;
-    }
-
-    float x, y, z;
-    if (m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION)
-    {
-        m_targets.getDestination(x, y, z);
-    }
-    else
-    {
-        m_caster->GetPosition(x, y, z);
-    }
-
-    float angle = unitTarget->GetAngle(x, y) + M_PI_F;
-    float horizontalSpeed = effect->EffectMiscValue * 0.1f;
-    float verticalSpeed = damage * 0.1f;
-    unitTarget->KnockBackWithAngle(angle, horizontalSpeed, verticalSpeed);
-}
-
-void Spell::EffectGravityPull(SpellEffectEntry const* effect)
-{
-    if (!unitTarget)
-    {
-        return;
-    }
-
-    float x, y, z;
-    if (m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION)
-    {
-        m_targets.getDestination(x, y, z);
-    }
-    else
-    {
-        m_caster->GetPosition(x, y, z);
-    }
-
-    float speed = float(effect->EffectMiscValue) * 0.15f;
-    float height = float(unitTarget->GetDistance(x, y, z) * 0.2f);
-
-    unitTarget->GetMotionMaster()->MoveJump(x, y, z, speed, height);
-}
-
-void Spell::EffectCreateTamedPet(SpellEffectEntry const* effect)
-{
-    if (!unitTarget || unitTarget->getClass() != CLASS_HUNTER)
-    {
-        return;
-    }
-
-    uint32 creatureEntry = effect->EffectMiscValue;
-
-    CreatureInfo const* cInfo = ObjectMgr::GetCreatureTemplate(creatureEntry);
-    if (creatureEntry && !cInfo)
-    {
-        sLog.outErrorDb("EffectCreateTamedPet: Creature entry %u not found for spell %u.", creatureEntry, m_spellInfo->Id);
-        return;
-    }
-
-    Pet* newTamedPet = new Pet(HUNTER_PET);
-    CreatureCreatePos pos(unitTarget, unitTarget->GetOrientation());
-
-    Map* map = unitTarget->GetMap();
-    uint32 petNumber = sObjectMgr.GeneratePetNumber();
-    if (!newTamedPet->Create(map->GenerateLocalLowGuid(HIGHGUID_PET), pos, cInfo, petNumber))
-    {
-        delete newTamedPet;
-        return;
-    }
-
-    newTamedPet->SetRespawnCoord(pos);
-
-    newTamedPet->SetOwnerGuid(unitTarget->GetObjectGuid());
-    newTamedPet->SetCreatorGuid(unitTarget->GetObjectGuid());
-    newTamedPet->SetUInt32Value(UNIT_NPC_FLAGS, UNIT_NPC_FLAG_NONE);
-    newTamedPet->setFaction(unitTarget->getFaction());
-    newTamedPet->SetUInt32Value(UNIT_FIELD_PET_NAME_TIMESTAMP, uint32(time(NULL)));
-    newTamedPet->SetUInt32Value(UNIT_CREATED_BY_SPELL, m_spellInfo->Id);
-
-    newTamedPet->GetCharmInfo()->SetPetNumber(petNumber, true);
-
-    if (unitTarget->IsPvP())
-    {
-        newTamedPet->SetPvP(true);
-    }
-
-    if (unitTarget->IsFFAPvP())
-    {
-        newTamedPet->SetFFAPvP(true);
-    }
-
-    newTamedPet->InitStatsForLevel(unitTarget->getLevel());
-    newTamedPet->InitPetCreateSpells();
-    newTamedPet->InitLevelupSpellsForLevel();
-    newTamedPet->InitTalentForLevel();
-
-    newTamedPet->SetByteFlag(UNIT_FIELD_BYTES_2, 2, UNIT_CAN_BE_RENAMED);
-    newTamedPet->SetByteFlag(UNIT_FIELD_BYTES_2, 2, UNIT_CAN_BE_ABANDONED);
-
-    newTamedPet->AIM_Initialize();
-
-    float x, y, z;
-    unitTarget->GetClosePoint(x, y, z, newTamedPet->GetObjectBoundingRadius());
-    newTamedPet->Relocate(x, y, z, unitTarget->GetOrientation());
-
-    map->Add((Creature*)newTamedPet);
-    m_caster->SetPet(newTamedPet);
-
-    newTamedPet->SavePetToDB(PET_SAVE_AS_CURRENT);
-    ((Player*)unitTarget)->PetSpellInitialize();
-}
-
